app/requirements.txt
fastapi==0.128.0
python-dotenv==1.2.1
sqlmodel==0.0.31
sqlalchemy==2.0.45
